# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PharmaRoute - Biomedical Research Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PaperStatusProperty = typeof PaperStatusProperty[keyof typeof PaperStatusProperty];


export const PaperStatusProperty = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Paper {
  id: string;
  title: string;
  /** @nullable */
  abstract?: string | null;
  authors?: string[];
  status: PaperStatusProperty;
  entitiesExtracted?: number;
  relationshipsExtracted?: number;
  createdAt: string;
  /** @nullable */
  processedAt?: string | null;
}

export interface PaperInput {
  title: string;
  abstract?: string;
  authors?: string[];
  content: string;
  sourceUrl?: string;
}

export interface PaperStatus {
  id: string;
  status: string;
  progress: number;
  currentStep?: string;
  entitiesFound?: number;
  relationshipsFound?: number;
}

export interface GraphStats {
  compounds: number;
  proteins: number;
  genes: number;
  diseases: number;
  pathways: number;
  papers: number;
  relationships: number;
  lastUpdated?: string;
}

export type GraphNodeType = typeof GraphNodeType[keyof typeof GraphNodeType];


export const GraphNodeType = {
  Compound: 'Compound',
  Protein: 'Protein',
  Gene: 'Gene',
  Disease: 'Disease',
  Pathway: 'Pathway',
  Paper: 'Paper',
} as const;

export type GraphNodeProperties = { [key: string]: unknown };

export interface GraphNode {
  id: string;
  name: string;
  type: GraphNodeType;
  properties?: GraphNodeProperties;
  connections?: number;
}

export interface PathQuery {
  fromId: string;
  toId: string;
  maxHops?: number;
}

export interface PathResult {
  found: boolean;
  path: GraphNode[];
  relationships?: string[];
  hops?: number;
  confidence?: number;
}

export interface Community {
  id: string;
  name: string;
  nodeCount: number;
  type: string;
  keyNodes?: string[];
  description?: string;
}

export interface Discovery {
  id: string;
  title: string;
  compound: string;
  disease: string;
  confidence: number;
  impactScore: number;
  biologicalPath: string[];
  evidenceCount: number;
  supportingPapers?: string[];
  explanation?: string;
  createdAt?: string;
}

export interface EvidenceStep {
  from: string;
  relationship: string;
  to: string;
  confidence: number;
  source?: string;
}

export interface DiscoveryDetail {
  id: string;
  title: string;
  compound: string;
  disease: string;
  confidence: number;
  impactScore: number;
  biologicalPath: string[];
  evidenceCount: number;
  supportingPapers?: string[];
  explanation: string;
  evidenceChain: EvidenceStep[];
  graphPath?: GraphNode[];
  createdAt?: string;
}

export interface DiscoveryRunInput {
  targetDisease?: string;
  minConfidence?: number;
  maxHops?: number;
}

export interface SearchQuery {
  query: string;
  nodeTypes?: string[];
  limit?: number;
}

export interface VoiceQuery {
  transcript: string;
}

export interface SearchResult {
  query: string;
  nodes: GraphNode[];
  discoveries: Discovery[];
  explanation: string;
  paths?: PathResult[];
}

export type ReportType = typeof ReportType[keyof typeof ReportType];


export const ReportType = {
  discovery: 'discovery',
  trends: 'trends',
  graph_analysis: 'graph_analysis',
  comprehensive: 'comprehensive',
} as const;

export type ResearchTrendDirection = typeof ResearchTrendDirection[keyof typeof ResearchTrendDirection];


export const ResearchTrendDirection = {
  rising: 'rising',
  stable: 'stable',
  declining: 'declining',
} as const;

export interface ResearchTrend {
  topic: string;
  score: number;
  paperCount: number;
  direction: ResearchTrendDirection;
  relatedTerms?: string[];
  description?: string;
}

export interface Report {
  id: string;
  title: string;
  type: ReportType;
  summary: string;
  discoveries?: Discovery[];
  trends?: ResearchTrend[];
  graphStats?: GraphStats;
  createdAt: string;
}

export type ReportInputType = typeof ReportInputType[keyof typeof ReportInputType];


export const ReportInputType = {
  discovery: 'discovery',
  trends: 'trends',
  graph_analysis: 'graph_analysis',
  comprehensive: 'comprehensive',
} as const;

export interface ReportInput {
  title: string;
  type: ReportInputType;
  targetDisease?: string;
}

export interface SeedResult {
  compounds: number;
  proteins: number;
  diseases: number;
  pathways: number;
  relationships: number;
  papers: number;
}

export type ListGraphNodesParams = {
type?: ListGraphNodesType;
limit?: number;
};

export type ListGraphNodesType = typeof ListGraphNodesType[keyof typeof ListGraphNodesType];


export const ListGraphNodesType = {
  Compound: 'Compound',
  Protein: 'Protein',
  Gene: 'Gene',
  Disease: 'Disease',
  Pathway: 'Pathway',
} as const;

export type ListDiscoveriesParams = {
minConfidence?: number;
limit?: number;
};