# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PharmaRoute - Biomedical Research Intelligence Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /papers:
    get:
      operationId: listPapers
      tags: [papers]
      summary: List all uploaded papers
      responses:
        "200":
          description: List of papers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Paper"
    post:
      operationId: uploadPaper
      tags: [papers]
      summary: Upload and process a scientific paper
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperInput"
      responses:
        "201":
          description: Paper uploaded and processing started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paper"
  /papers/{id}:
    get:
      operationId: getPaper
      tags: [papers]
      summary: Get paper by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Paper details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paper"
        "404":
          description: Not found
  /papers/{id}/status:
    get:
      operationId: getPaperStatus
      tags: [papers]
      summary: Get processing status of a paper
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Processing status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperStatus"
...
  /graph/nodes:
    get:
      operationId: listGraphNodes
      tags: [graph]
      summary: List nodes in the graph
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [Compound, Protein, Gene, Disease, Pathway]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of graph nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GraphNode"
  /graph/paths:
    post:
      operationId: findGraphPath
      tags: [graph]
      summary: Find shortest path between two nodes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PathQuery"
      responses:
        "200":
          description: Path result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PathResult"
...
  /discoveries:
    get:
      operationId: listDiscoveries
      tags: [discoveries]
      summary: List drug repurposing discoveries
      parameters:
        - name: minConfidence
          in: query
          schema:
            type: number
            default: 0.5
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of discoveries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Discovery"
  /discoveries/{id}:
    get:
      operationId: getDiscovery
      tags: [discoveries]
      summary: Get discovery details with evidence chain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Discovery detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoveryDetail"
  /discoveries/run:
    post:
      operationId: runDiscoveryEngine
      tags: [discoveries]
      summary: Trigger the drug repurposing discovery engine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscoveryRunInput"
      responses:
        "200":
          description: Discovery run results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Discovery"
...
components:
  schemas:
...
    Paper:
      type: object
      required: [id, title, status, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        abstract:
          type: ["string", "null"]
        authors:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        entitiesExtracted:
          type: integer
        relationshipsExtracted:
          type: integer
        createdAt:
          type: string
        processedAt:
          type: ["string", "null"]
    PaperInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        abstract:
          type: string
        authors:
          type: array
          items:
            type: string
        content:
          type: string
        sourceUrl:
          type: string
    PaperStatus:
      type: object
      required: [id, status, progress]
      properties:
        id:
          type: string
        status:
          type: string
        progress:
          type: integer
        currentStep:
          type: string
        entitiesFound:
          type: integer
        relationshipsFound:
          type: integer
...
    GraphNode:
      type: object
      required: [id, name, type]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [Compound, Protein, Gene, Disease, Pathway, Paper]
        properties:
          type: object
          additionalProperties: true
        connections:
          type: integer
    PathQuery:
      type: object
      required: [fromId, toId]
      properties:
        fromId:
          type: string
        toId:
          type: string
        maxHops:
          type: integer
          default: 5
    PathResult:
      type: object
      required: [found, path]
      properties:
        found:
          type: boolean
        path:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        relationships:
          type: array
          items:
            type: string
        hops:
          type: integer
        confidence:
          type: number
...
    Discovery:
      type: object
      required: [id, title, compound, disease, confidence, impactScore, biologicalPath, evidenceCount]
      properties:
        id:
          type: string
        title:
          type: string
        compound:
          type: string
        disease:
          type: string
        confidence:
          type: number
        impactScore:
          type: number
        biologicalPath:
          type: array
          items:
            type: string
        evidenceCount:
          type: integer
        supportingPapers:
          type: array
          items:
            type: string
        explanation:
          type: string
        createdAt:
          type: string
    DiscoveryDetail:
      type: object
      required: [id, title, compound, disease, confidence, impactScore, biologicalPath, evidenceCount, evidenceChain, explanation]
      properties:
        id:
          type: string
        title:
          type: string
        compound:
          type: string
        disease:
          type: string
        confidence:
          type: number
        impactScore:
          type: number
        biologicalPath:
          type: array
          items:
            type: string
        evidenceCount:
          type: integer
        supportingPapers:
          type: array
          items:
            type: string
        explanation:
          type: string
        evidenceChain:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceStep"
        graphPath:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        createdAt:
          type: string
    EvidenceStep:
      type: object
      required: [from, relationship, to, confidence]
      properties:
        from:
          type: string
        relationship:
          type: string
        to:
          type: string
        confidence:
          type: number
        source:
          type: string
    DiscoveryRunInput:
      type: object
      properties:
        targetDisease:
          type: string
        minConfidence:
          type: number
          default: 0.5
        maxHops:
          type: integer
          default: 4
...