# lib/api-zod/src/generated/types/discovery.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PharmaRoute - Biomedical Research Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */

export interface Discovery {
  id: string;
  title: string;
  compound: string;
  disease: string;
  confidence: number;
  impactScore: number;
  biologicalPath: string[];
  evidenceCount: number;
  supportingPapers?: string[];
  explanation?: string;
  createdAt?: string;
}