# lib/api-zod/src/generated/types/discoveryDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PharmaRoute - Biomedical Research Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
import type { EvidenceStep } from './evidenceStep';
import type { GraphNode } from './graphNode';

export interface DiscoveryDetail {
  id: string;
  title: string;
  compound: string;
  disease: string;
  confidence: number;
  impactScore: number;
  biologicalPath: string[];
  evidenceCount: number;
  supportingPapers?: string[];
  explanation: string;
  evidenceChain: EvidenceStep[];
  graphPath?: GraphNode[];
  createdAt?: string;
}