# lib/api-zod/src/generated/types/graphNode.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PharmaRoute - Biomedical Research Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
import type { GraphNodeProperties } from './graphNodeProperties';
import type { GraphNodeType } from './graphNodeType';

export interface GraphNode {
  id: string;
  name: string;
  type: GraphNodeType;
  properties?: GraphNodeProperties;
  connections?: number;
}