# lib/api-zod/src/generated/types/pathResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PharmaRoute - Biomedical Research Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
import type { GraphNode } from './graphNode';

export interface PathResult {
  found: boolean;
  path: GraphNode[];
  relationships?: string[];
  hops?: number;
  confidence?: number;
}